cmd: pull closure out into its own func
Also introduce a config.go.
